requirements.txt
langchain
langchain-community
langchain-pinecone
langchain-google-genai
pinecone-client
sentence-transformers
pypdf
python-dotenv
